fix: `installQuartoExtensionSource` to keep "source" for install/update
